pyproject.toml
[project]
name = "mimic-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "db-dtypes>=1.4.2",
    "google-cloud-bigquery>=3.31.0",
    "ipykernel>=6.29.5",
    "polars>=1.26.0",
    "pydantic>=2.11.2",
    "pyyaml>=6.0.2",
    "streamlit>=1.44.1",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.4",
]
